% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c1_labels}
\alias{c1_labels}
\title{Clustering Labels of Simulated Objects}
\format{
An integer vector.
\describe{
  \item{c1_labels}{First clustering result to compare}
}
}
\description{
A manually generated clustering labels.
}
