% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coords}
\alias{coords}
\title{Coordinates of Simulated Objects}
\format{
A data frame with 2 columns:
\describe{
  \item{X_coord}{x-coordinate}
  \item{Y_coord}{y-coordinate}
}
}
\description{
A simulated dataset containing spatial coordinates.
}
