% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_sg_pairs_int}
\alias{generate_sg_pairs_int}
\title{Generate object pairs (sg pairs)}
\arguments{
\item{c_labels}{An integer vector of clustering labels.}

\item{r_labels}{An integer vector of reference labels.}
}
\value{
An integer matrix with two columns, where each row is a pair of object indices.
}
\description{
This function identifies all object pairs that belong to the same cluster in the clustering partition
but different groups in the reference partition
}
\examples{
c_labels <- c(1,1,2,2,2,3,3)
r_labels <- c(1,1,1,2,2,3,1)
generate_sg_pairs_int(c_labels, r_labels)

}
