% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{true_labels}
\alias{true_labels}
\title{True Labels of Simulated Objects}
\format{
An integer vector.
\describe{
  \item{true_labels}{Ground-truth labels of 160 spots}
}
}
\description{
A manually generated true labels.
}
