% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{spARI_example_data}
\alias{spARI_example_data}
\title{Example data for spARI}
\format{
An environment containing 4 objects:
\describe{
  \item{true_labels}{Numeric or factor vector of length 160.}
  \item{c1_labels}{Numeric or factor vector of length 160.}
  \item{c2_labels}{Numeric or factor vector of length 160.}
  \item{coords}{Numeric matrix of size 160x2.}
}
}
\usage{
data(spARI_example_data)
}
\description{
This dataset includes simulated spatial clustering results for demonstration purposes.
It contains the following objects:
\itemize{
  \item \code{true_labels}: Ground-truth labels of 160 spots.
  \item \code{c1_labels}: First clustering result to compare.
  \item \code{c2_labels}: Second clustering result to compare.
  \item \code{coords}: Spatial coordinates (matrix with 2 columns).
}
}
\keyword{datasets}
